Add OwnerReference to operator created objects
